backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
aiosqlite>=0.19.0
greenlet>=3.0.0
# Note: rclpy is usually installed via ROS 2 system packages, 
# but we list it here for completeness if using a venv that can see system packages.
# In a pure pip env without ROS sourced, this might fail, but we assume ROS environment.
